old/iot-hub/requirements.txt
paho-mqtt==2.1.0
psycopg2-binary==2.9.10
requests==2.32.3
